ai_agent_project/requirements.txt
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
chromadb>=0.4.0
termcolor>=2.3.0
colorama>=0.4.6
pyyaml>=6.0
tenacity>=8.2.3
google-genai>=0.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
sse-starlette>=0.0.0
